requirements.txt
Django==3.2.6
django-cors-headers==3.7.0
django-mysql==3.12.0
djangorestframework==3.12.4
mysqlclient==2.0.3
PyJWT==2.1.0
Faker==8.12.1
django-redis==4.12.1
stripe==2.56.0